Merge TreeSummarizedExperiment objects with the same assay types together

'mergeExperiments' takes a list of TreeSummarizedExperiment objects with the same assays but different samples, and combines them into a single TreeSummarizedExperiment object.

Usage

mergeExperiments(merge_list)

Arguments

merge_list

List of TreeSummarizedExperiment objects: to be merged into a single TreeSummarizedExperiment object

Value

TreeSummarizedExperiment object with multiple samples

Details

The assays contained in the TreeSummarizedExperiments to be merged must be of the same type (i.e. have the same name) and be in the same order if there are multiple assays.

See also

Examples

if (FALSE) { # \dontrun{
if(interactive()){
 fpath <- file.path(system.file("extdata",
                                package = "parkinsonsMetagenomicData"),
                    "sample_experiment_list.Rds")
 sample_experiment_list <- readRDS(fpath)

 mergeExperiments(sample_experiment_list)
 }
} # }