Parse HUMAnN output for a single sample as a TreeSummarizedExperiment object

'parse_humann' reads a file obtained from running HUMAnN on a single sample. This file is parsed into a TreeSummarizedExperiment object.

Usage

parse_humann(sample_id, file_path, data_type)

Arguments

sample_id

String: A sample identifier

file_path

String: Path to a locally stored HUMAnN output file in gzipped TSV format

data_type

String: The type of HUMAnN output file to be parsed, as found in output_file_types("tool", "humann")

Value

A TreeSummarizedExperiment object with process metadata, row names, column names, and relevant assays.

Details

This function does not integrate sample metadata as column data. The provided sample_id is used as the column name for assays within the TreeSummarizedExperiment object and is intended to be used for integration of sample metadata.

See also

Examples

if (FALSE) { # \dontrun{
if(interactive()){
 #EXAMPLE1
 }
} # }